refactor(#276) repositoryディレクトリを移動
