+++ backend/internal/adapter/controller/article_test.go
@@ -9,9 +9,9 @@ import (
 	"github.com/bufbuild/connect-go"
 	"github.com/google/uuid"
 	"github.com/morning-night-guild/platform/internal/adapter/controller"
-	"github.com/morning-night-guild/platform/internal/model"
-	"github.com/morning-night-guild/platform/internal/model/article"
-	me "github.com/morning-night-guild/platform/internal/model/errors"
+	"github.com/morning-night-guild/platform/internal/domain/model"
+	"github.com/morning-night-guild/platform/internal/domain/model/article"
+	me "github.com/morning-night-guild/platform/internal/domain/model/errors"
 	"github.com/morning-night-guild/platform/internal/usecase"
 	"github.com/morning-night-guild/platform/internal/usecase/port"
 	articlev1 "github.com/morning-night-guild/platform/pkg/connect/proto/article/v1"
