+++ backend/internal/adapter/gateway/article_test.go
@@ -9,9 +9,9 @@ import (
 	"github.com/google/uuid"
 	_ "github.com/mattn/go-sqlite3"
 	"github.com/morning-night-guild/platform/internal/adapter/gateway"
-	"github.com/morning-night-guild/platform/internal/model"
-	"github.com/morning-night-guild/platform/internal/model/article"
-	"github.com/morning-night-guild/platform/internal/usecase/repository"
+	"github.com/morning-night-guild/platform/internal/domain/model"
+	"github.com/morning-night-guild/platform/internal/domain/model/article"
+	"github.com/morning-night-guild/platform/internal/domain/repository"
 	"github.com/morning-night-guild/platform/pkg/ent"
 	"github.com/morning-night-guild/platform/pkg/ent/articletag"
 	"github.com/morning-night-guild/platform/pkg/ent/enttest"
