+++ backend/internal/domain/model/article/description.go
