+++ backend/internal/domain/model/article/tag_list.go
@@ -3,7 +3,7 @@ package article
 import (
 	"fmt"
 
-	"github.com/morning-night-guild/platform/internal/model/errors"
+	"github.com/morning-night-guild/platform/internal/domain/model/errors"
 )
 
 // Tag タグリスト.
