+++ backend/internal/usecase/interactor/article/list.go
@@ -3,8 +3,8 @@ package article
 import (
 	"context"
 
+	"github.com/morning-night-guild/platform/internal/domain/repository"
 	"github.com/morning-night-guild/platform/internal/usecase/port"
-	"github.com/morning-night-guild/platform/internal/usecase/repository"
 )
 
 var _ port.ListArticle = (*ListInteractor)(nil)
