+++ backend/internal/usecase/mock/article.go
@@ -3,8 +3,8 @@ package mock
 import (
 	"context"
 
-	"github.com/morning-night-guild/platform/internal/model"
-	"github.com/morning-night-guild/platform/internal/usecase/repository"
+	"github.com/morning-night-guild/platform/internal/domain/model"
+	"github.com/morning-night-guild/platform/internal/domain/repository"
 )
 
 var _ repository.Article = (*Article)(nil)
